Renamed to match new style and lots of changes
I'm trying to add in a lot of debugging stuff as I work through the issues. Building a game bottom-up (or sideways-up) is much harder, but still fun! :)

The three Libraries are similar, but they each function a little differently, so I'm not sure if I can make a generic version or not... still pondering.