Retcon, gogogo!
Sooo... I ended up looking over the code and deciding to change a bunch of stuff. That means learning how to branch (woo!) as I overhaul it. Exits are turning into just another kind of Action and I'm slowly building up the story for the outside of the house. This should be a fun week! :)