Room 1 Basic Testing Outline
I wrote out the items that are in Room 1 that need to be tested and the results of my first round of testing. Next, I will tackle fixing the bugs in this room. Once that is done I will merge the branch back into master and open a new branch for the next room.